+++ Controllers/CreateTemplate.php
@@ -1,15 +1,15 @@
 <?php
 
-namespace Leantime\Plugins\DefaultTicketTemplate\Controllers;
+namespace Leantime\Plugins\TicketTemplate\Controllers;
 
 use Leantime\Core\Controller;
 use Leantime\Core\Frontcontroller;
-use Leantime\Plugins\DefaultTicketTemplate\Repository\DefaultTicketTemplateRepository;
+use Leantime\Plugins\TicketTemplate\Repository\TicketTemplateRepository;
 use Symfony\Component\HttpFoundation\RedirectResponse;
 use Symfony\Component\HttpFoundation\Response;
 
 /**
- * Create template Controller for DefaultTicketTemplate plugin
+ * Create template Controller for TicketTemplate plugin
  *
  * @package    leantime
  * @subpackage plugins
@@ -28,7 +28,7 @@ class CreateTemplate extends Controller
         // without an extra call to readIni().
         $this->language->readIni();
 
-        return $this->tpl->display('defaultTicketTemplate.createTemplate');
+        return $this->tpl->display('TicketTemplate.createTemplate');
     }
 
     /**
@@ -41,14 +41,14 @@ class CreateTemplate extends Controller
     public function post(array $params): RedirectResponse
     {
         if (isset($params['title']) && isset($params['content'])) {
-            $defaultTicketTemplateRepository = app()->make(DefaultTicketTemplateRepository::class);
-            $defaultTicketTemplateRepository->addTemplate($params['title'], $params['content']);
+            $ticketTemplateRepository = app()->make(TicketTemplateRepository::class);
+            $ticketTemplateRepository->addTemplate($params['title'], $params['content']);
 
             $this->tpl->setNotification(__('tickettemplate.create.success_message'), 'success');
         } else {
             $this->tpl->setNotification(__('tickettemplate.create.failed_message'), 'error');
         }
 
-        return Frontcontroller::redirect(BASE_URL . '/DefaultTicketTemplate/listTemplates');
+        return Frontcontroller::redirect(BASE_URL . '/TicketTemplate/listTemplates');
     }
 }