+++ Controllers/DeleteTemplate.php
@@ -1,16 +1,16 @@
 <?php
 
-namespace Leantime\Plugins\DefaultTicketTemplate\Controllers;
+namespace Leantime\Plugins\TicketTemplate\Controllers;
 
 use Leantime\Core\Controller;
 use Leantime\Core\Frontcontroller;
 use Leantime\Domain\Auth\Models\Roles;
 use Leantime\Domain\Auth\Services\Auth;
-use Leantime\Plugins\DefaultTicketTemplate\Repository\DefaultTicketTemplateRepository;
+use Leantime\Plugins\TicketTemplate\Repository\TicketTemplateRepository;
 use Symfony\Component\HttpFoundation\Response;
 
 /**
- * Create template Controller for DefaultTicketTemplate plugin
+ * Create template Controller for TicketTemplate plugin
  *
  * @package    leantime
  * @subpackage plugins
@@ -31,21 +31,21 @@ class DeleteTemplate extends Controller
 
         Auth::authOrRedirect([Roles::$owner, Roles::$admin], true);
 
-        $defaultTicketTemplateRepository = app()->make(DefaultTicketTemplateRepository::class);
+        $ticketTemplateRepository = app()->make(TicketTemplateRepository::class);
 
         if (isset($_GET['id']) === true) {
             $id = (int) ($_GET['id']);
 
             if (isset($_POST['del']) === true) {
-                $defaultTicketTemplateRepository->deleteTemplate($id);
+                $ticketTemplateRepository->deleteTemplate($id);
 
                 $this->tpl->setNotification($this->language->__('tickettemplate.delete.success_message'), "success");
 
-                return Frontcontroller::redirect(BASE_URL . "/DefaultTicketTemplate/listTemplates");
+                return Frontcontroller::redirect(BASE_URL . "/TicketTemplate/listTemplates");
             }
 
             //Assign template.
-            $template = $defaultTicketTemplateRepository->getTemplateById($id);
+            $template = $ticketTemplateRepository->getTemplateById($id);
 
             if ($template === false) {
                 return $this->tpl->display('errors.error403', responseCode: 403);
@@ -53,7 +53,7 @@ class DeleteTemplate extends Controller
 
             $this->tpl->assign('template', $template);
 
-            return $this->tpl->display('defaultTicketTemplate.deleteTemplate');
+            return $this->tpl->display('ticketTemplate.deleteTemplate');
         } else {
             return $this->tpl->display('errors.error403', responseCode: 403);
         }