+++ register.php
@@ -3,14 +3,14 @@
 use Leantime\Core\Application;
 use Leantime\Core\Events;
 use Leantime\Domain\Tickets\Models\Tickets;
-use Leantime\Plugins\DefaultTicketTemplate\Services\DefaultTicketTemplate;
+use Leantime\Plugins\TicketTemplate\Services\TicketTemplate;
 
 app()->resolving(Tickets::class, function (Tickets $ticket, Application $application) {
-    $application->make(DefaultTicketTemplate::class)->processNewTicket($ticket);
+    $application->make(TicketTemplate::class)->processNewTicket($ticket);
 });
 
 // Add plugin translation file,
-// see @Leantime\Plugins\DefaultTicketTemplate\Controllers:get(),
+// see @Leantime\Plugins\TicketTemplate\Controllers:get(),
 // for how we ensure these are read.
 Events::add_filter_listener(
     'leantime.core.language.readIni.language_files',